Created hook component to test in tC
